New features in the services

+++ src/main/kotlin/entities/Socio.kt
@@ -3,12 +3,12 @@ package entities
 import enumerations.DescricaoSocio
 
 data class Socio(
-    var id : Long,
-    var nomeSocio : String,
-    var enderecoSocio : String,
-    var cepSocio : Long,
-    var emailSocio : String,
-    var senha : String,
-    var tipoSocio : DescricaoSocio
+    var id: Long,
+    var nomeSocio: String,
+    var enderecoSocio: String,
+    var cepSocio: Int,
+    var emailSocio: String,
+    var senha: String,
+    var tipoSocio: DescricaoSocio
 )
